Split game settings
